third_party/WebKit/LayoutTests/fast/serviceworker/serviceworker-interface-expected.txt

This is a testharness.js-based test.
FAIL ServiceWorker interface assert_equals: ServiceWorker should extend Worker expected object "[object Worker]" but got object "[object EventTarget]"
Harness: the test ran to completion.
